beyond, infinit-user: handle duplicate user with same public_key
beyond response to PUT <user> is now 200 if the user exists with the same
public_key.
It response 409 if the user exists but the public_key is different.

infinit-user output a friendly message if the server already has the user with
the same public_key associated instead of a fatal error.

* beyond/src/bottle.py: handle duplicate user with same key.
* beyond/tests/full-flow: add 2 unit tests.
* beyond/tests/utils.py: add mefy2.
* bin/main.hh.tmpl: add an if statement.