main.hh: address some of the warnings
We have tons of warnings in main.hh.tmpl because the API of Serializer
was updated, but not its uses here.  Make a partial migration to the
new API, but not a full one, as some of the calls appear to still need
the name.

Anyway this is scheduled to be soon replaced by something completely
different.