tests: adjust cli tests
After the current overhaul of CLI, --push, being a Boolean option,
will take an optional argument: true/false.  Unfortunately, whatever
argument behind it will be considered attached to --push, so

    infinit-drive --create --as bob --network n --volume v --push d

will fail because `d` is passed to `--push`.  It was confirmed with
Quentin Hocquet that this is the desired behavior.  So change the
tests to run

    infinit-drive --create d --as bob --network n --volume v --push

instead.