cli: Volume: do not register default values
As is currently implemented, the mount options make no difference
between default values, and user values.  So they are all exported
alike.  This is not the desired behavior.

Instead of using boost::optional everywhere, which makes the code
tedious, and the --help messages less interesting (as they cannot show
the default values), we want das.cli to provide us with values that
can tell whether they are from the user or not.  That's
das::cli::Defaulted.

Use them.