cli.Doctor: introduce a structure that deals with output features
Currently doctor still has a couple of global variables (!) used.
They were in the legacy implementation, and were kept during the
migration to make it easier.

Before dealing with these globals (one being about how we output the
results), let's group the output stream and whether we use colors in a
single object, to avoid functions with a bazillion arguments.