cli.Doctor: tidy
Note that GCC on Jessie fails to make static local variables visible
to lambdas.  So banished_log_level could not be made a local static
variable.  But anyway, it's very dubious that we have any interest in
building such a string, as anyway std::string's operator== is
specialized for comparison with C strings.