overlay.tests: clean up
I am having a hard time understanding the logic of the tests, because
the separation Kelips/Kouncil is not that clear.  So instead of
checking for boost::bad_any_cast (because we try to convert an Null
json object because we use operator[] which creates an empty object
rather than saying that the key does not exist...) to see that a key
does not exist, decide for the key depending on the overlay type.