overlay: fix failures on Windows
These failures are actually due to the change of the eviction delay,
which became way too low and prevented reconnections.

Let's leave the eviction delay to its default value everywhere, and
change it only where we want it to be short.  Which means adding
modifiers.  Let's use this as an optortunity to use a better type for
the eviction delay.

So, also clean up, but don't reenable the eviction test just yet: we
still have a deadlock somewhere.