prometheus: fix the test on Windows
By not running it at all there.  The problem is that we do not use
Boost.Test and we provide our own main.  But because of elle/test.hh
we need to link against Boost.Test.  This is ok on Unix since the lib
is shared, but it fails on Windows (two `main`s) because the lib is
static.

Eventually we should use Boost.Test's main, but in meanwhile, since
anyway prometheus is not supported on Windows, just skip the test
there.