memo.log: don't start a new log for network, rename the current one
If we just start a new one, we lose everything that was logged before.
Also, after discussing with Quentin Hocquet, it is prefered to start
new log files rather than appending to existing ones (which naturally
follows from the desire to rename).

Also, to avoid name clashes, we now use the time stamp.